docker/vertex-training/requirements.txt
# Vertex AI training container dependencies (minimal)
pandas>=2.0,<2.2
numpy>=1.24,<1.26
scikit-learn>=1.3,<1.5
joblib>=1.3,<2.0

google-cloud-bigquery>=3.10,<4.0
google-cloud-storage>=2.13,<3.0

# BigQuery <-> pandas helpers
db-dtypes>=1.1,<2.0
pyarrow>=12.0,<14.0

